update perf test script
